Create email_notificationSystem
Simple email notification system using a Queue. I used a LinkedList to represent the queue of email notifications, I implemented methods to add new email notifications to the queue and process them in FIFO order. Finally, I displayed the current queue status and the next email notification to be processed.